requirements.txt
gym >= 0.26
numpy==1.24.3
# jax==0.4.13
# jax==0.4.19
# distrax==0.1.2
# flax==0.7.0
ml_collections >= 0.1.0
tqdm >= 4.60.0
# chex==0.1.6
# optax==0.1.5
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.12.14
# tensorflow==2.13.0
# tensorflow_probability==0.21
# tensorflow_hub
# tensorflow_text
einops >= 0.6.1
imageio >= 2.31.1
moviepy >= 1.0.3
# orbax
matplotlib
pyquaternion
opencv-python
opencv-contrib-python
funcsigs


distrax==0.1.2
flax==0.7.0
einops==0.6.1
# wandb==0.15.5
wandb
jax==0.4.11
jaxlib==0.4.11
optax==0.1.5
diffusers==0.18.2
ml-dtypes==0.2.0
ipdb 

tensorflow[and-cuda] 
tensorflow_datasets>=4.9.2 
# tensorflow_probability==0.21
# tensorflow_hub
# tensorflow_text
tqdm-multiprocess==0.0.11 
pre-commit==3.3.3
typing_extensions